Switch from gometalinter to golangci-lint
The maintainer of gometalinter has deprecated its tool and advise now to
use golangci-lint instead. We can't use the latest version, because it's
not compatible with the current version of Go.

Use the shell script installer instead of `go get` as the authors of the
tool say in the README. It's faster and more reliable to download
directly the binary that build it in.